refactor the code to single file
