+++ .idea/workspace.xml
@@ -1,0 +1,103 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="BranchesTreeState">
+    <expand>
+      <path>
+        <item name="ROOT" type="e8cecc67:BranchNodeDescriptor" />
+        <item name="LOCAL_ROOT" type="e8cecc67:BranchNodeDescriptor" />
+      </path>
+      <path>
+        <item name="ROOT" type="e8cecc67:BranchNodeDescriptor" />
+        <item name="REMOTE_ROOT" type="e8cecc67:BranchNodeDescriptor" />
+      </path>
+      <path>
+        <item name="ROOT" type="e8cecc67:BranchNodeDescriptor" />
+        <item name="REMOTE_ROOT" type="e8cecc67:BranchNodeDescriptor" />
+        <item name="GROUP_NODE:origin" type="e8cecc67:BranchNodeDescriptor" />
+      </path>
+    </expand>
+    <select />
+  </component>
+  <component name="ChangeListManager">
+    <list default="true" id="9f8e17ae-e3e8-4587-b332-48df83b3c26b" name="Default Changelist" comment="">
+      <change beforePath="$PROJECT_DIR$/common/header.php" beforeDir="false" afterPath="$PROJECT_DIR$/common/header.php" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/func/api.php" beforeDir="false" afterPath="$PROJECT_DIR$/func/api.php" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/func/db.php" beforeDir="false" afterPath="$PROJECT_DIR$/func/db.php" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/func/func.php" beforeDir="false" afterPath="$PROJECT_DIR$/func/func.php" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/func/member_func.php" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/func/message_func.php" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/func/posting_func.php" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/func/social_func.php" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/member/base_info.php" beforeDir="false" afterPath="$PROJECT_DIR$/member/base_info.php" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/member/contract.php" beforeDir="false" afterPath="$PROJECT_DIR$/member/contract.php" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/member/hot_posts.php" beforeDir="false" afterPath="$PROJECT_DIR$/member/hot_posts.php" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/member/inbox.php" beforeDir="false" afterPath="$PROJECT_DIR$/member/inbox.php" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/member/index.php" beforeDir="false" afterPath="$PROJECT_DIR$/member/index.php" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/member/login.php" beforeDir="false" afterPath="$PROJECT_DIR$/member/login.php" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/member/message.php" beforeDir="false" afterPath="$PROJECT_DIR$/member/message.php" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/member/nav.php" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/member/posting.php" beforeDir="false" afterPath="$PROJECT_DIR$/member/posting.php" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/member/posting_templ.php" beforeDir="false" afterPath="$PROJECT_DIR$/member/posting_templ.php" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/member/sentbox.php" beforeDir="false" afterPath="$PROJECT_DIR$/member/sentbox.php" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/member/social.php" beforeDir="false" afterPath="$PROJECT_DIR$/member/social.php" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/static/functions.js" beforeDir="false" afterPath="$PROJECT_DIR$/static/functions.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/static/message.js" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/static/posting.js" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/static/social.js" beforeDir="false" />
+    </list>
+    <option name="SHOW_DIALOG" value="false" />
+    <option name="HIGHLIGHT_CONFLICTS" value="true" />
+    <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
+    <option name="LAST_RESOLUTION" value="IGNORE" />
+  </component>
+  <component name="ComposerSettings">
+    <execution />
+  </component>
+  <component name="Git.Settings">
+    <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
+  </component>
+  <component name="ProjectId" id="1l5daKfYFXLuCgfQN4iLPdkNQiX" />
+  <component name="ProjectViewState">
+    <option name="hideEmptyMiddlePackages" value="true" />
+    <option name="showLibraryContents" value="true" />
+  </component>
+  <component name="PropertiesComponent">
+    <property name="RunOnceActivity.OpenProjectViewOnStart" value="true" />
+    <property name="WebServerToolWindowFactoryState" value="false" />
+    <property name="last_opened_file_path" value="$PROJECT_DIR$" />
+    <property name="nodejs_package_manager_path" value="npm" />
+    <property name="vue.rearranger.settings.migration" value="true" />
+  </component>
+  <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
+  <component name="TaskManager">
+    <task active="true" id="Default" summary="Default task">
+      <changelist id="9f8e17ae-e3e8-4587-b332-48df83b3c26b" name="Default Changelist" comment="" />
+      <created>1606890947698</created>
+      <option name="number" value="Default" />
+      <option name="presentableId" value="Default" />
+      <updated>1606890947698</updated>
+      <workItem from="1606890949203" duration="2439000" />
+    </task>
+    <servers />
+  </component>
+  <component name="TypeScriptGeneratedFilesManager">
+    <option name="version" value="3" />
+  </component>
+  <component name="Vcs.Log.Tabs.Properties">
+    <option name="TAB_STATES">
+      <map>
+        <entry key="MAIN">
+          <value>
+            <State />
+          </value>
+        </entry>
+      </map>
+    </option>
+  </component>
+  <component name="WindowStateProjectService">
+    <state x="500" y="160" key="#com.intellij.refactoring.safeDelete.UnsafeUsagesDialog" timestamp="1606892874496">
+      <screen x="0" y="0" width="1536" height="824" />
+    </state>
+    <state x="500" y="160" key="#com.intellij.refactoring.safeDelete.UnsafeUsagesDialog/0.0.1536.824@0.0.1536.824" timestamp="1606892874495" />
+  </component>
+</project>
