+++ func/db.php
@@ -4,8 +4,8 @@
 try {
 
     $user = 'root';
-    $pass = 'Alex112524';
-    $host = '127.0.0.1:3306';
+    $pass = 'hkc610787';
+    $host = '127.0.0.1:3307';
     $dbname = 'condo_system';
